Backend/requirements.txt
fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
python_jose==3.5.0
pytz==2025.2
SQLAlchemy==2.0.41
uvicorn
email-validator
python-dotenv
bcrypt
google-generativeai         
transformers                
torch         
